+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Initial support for extra damage from vulnerabilities. [#1479](https://github.com/dmdorman/hero6e-foundryvtt/issues/1479)
 - Conditional defenses and vulnerabilities will attempt to automatically determine if it applies (uses SFX and Description). GM still has to confirm conditional effects.
 - Fixed Flash, which now has a unique icon per sense group. [#1486](https://github.com/dmdorman/hero6e-foundryvtt/issues/1486)
+- Support the 1/2 endurance version of COSTSEND. [#1497](https://github.com/dmdorman/hero6e-foundryvtt/issues/1497)
 
 ## Version 4.0.6